+++ src/main/java/subway/domain/fare/RouteFareStrategy.java
@@ -8,7 +8,7 @@ import subway.domain.subway.Subway;
 public class RouteFareStrategy implements FareStrategy {
 
     @Override
-    public long calculateFare(final long fare, final Passenger passenger, final Subway subway) {
+    public double calculateFare(final double fare, final Passenger passenger, final Subway subway) {
         final List<PathSection> sections = subway.findShortestPathSections(passenger.getStart(), passenger.getEnd());
         final int highestFareOfLine = findHighestFareOfLine(sections);
         return fare + highestFareOfLine;
